✨ refactor(document-form): refactor sign component and extract to separate file
