+++ app/components/document/Dialog/Sign.vue
@@ -1,20 +1,10 @@
 <script setup lang="ts">
-import { z } from 'zod';
+import { useForwardPropsEmits } from 'radix-vue';
+import type { SignEmits } from '@/components/document/form/Sign.vue';
 
-type SignEmits = {
-  sign: [];
-};
-const emits = defineEmits<SignEmits>();
-
-const signValue = defineModel<string>();
-
-const schema = z.object({
-  file: z.string().optional(),
-});
-
-function handleSubmit() {
-  emits('sign');
-};
+const props = defineProps<{ modelValue: string }>();
+const emits = defineEmits<SignEmits & { 'update:modelValue': [value: string] }>();
+const forwarded = useForwardPropsEmits(props, emits);
 </script>
 
 <template>
@@ -27,43 +17,7 @@ function handleSubmit() {
         </ShadcnDialogDescription>
       </ShadcnDialogHeader>
 
-      <ShadcnTabs
-        default-value="file"
-      >
-        <ShadcnTabsList class="grid w-full grid-cols-2">
-          <ShadcnTabsTrigger value="file">
-            File
-          </ShadcnTabsTrigger>
-          <ShadcnTabsTrigger value="canvas">
-            Canvas
-          </ShadcnTabsTrigger>
-        </ShadcnTabsList>
-        <ShadcnTabsContent value="file">
-          <ShadcnAutoForm
-            :schema
-            @submit="handleSubmit"
-          >
-            <template #file="slotProps">
-              <ShadcnAutoFormFieldFile
-                v-model="signValue"
-                v-bind="slotProps"
-              />
-            </template>
-
-            <div class="mt-2 flex justify-end">
-              <ShadcnButton type="submit">
-                Save
-              </ShadcnButton>
-            </div>
-          </ShadcnAutoForm>
-        </ShadcnTabsContent>
-        <ShadcnTabsContent value="canvas">
-          <BaseInputSign
-            v-model:url="signValue"
-            @save="handleSubmit"
-          />
-        </ShadcnTabsContent>
-      </ShadcnTabs>
+      <DocumentFormSign v-bind="forwarded" />
     </ShadcnDialogContent>
   </ShadcnDialog>
 </template>
